change integration test from 2.2 to 2.0.5
